feat(flake.nix): Refactor flake.nix and related files
This commit refactors the flake.nix file and related files. The changes
include moving the definition of `pkgs`, `currentUsername`, and
`currentHomeDirectory` to a different location within the flake.nix file.
The `nixOSRebuild` script is also moved accordingly.

In addition, the `home-manager/tmux/default.nix` and `home.nix` files are
updated to remove unnecessary inputs. This simplifies the structure of
these files and makes them easier to understand and maintain.