feat(flake.nix, home.nix): Add local rebuild option
This commit introduces a new option to the `update-os` and `update-hm`
scripts. The `--local` or `-l` option allows the user to rebuild the
system or home-manager configuration using the local flake.nix file,
instead of fetching it from the remote repository. This is useful for
testing changes locally before pushing them to the remote repository.