feat: Add new packages and update keybinds
This commit includes two major changes. First, it adds new packages
to the home-manager configuration: gh-copilot, gh-dash, gh-poi, and gh-s.

Second, it updates the keybindings in the Zellij configuration. The
"Ctrl [" "Enter" "Esc" "Ctrl c" sequence now includes "i", which
switches to the "normal" mode. This change is applied in the "search",
"scroll", and "tmux" modes.